Move services logic to handlers
