+++ src/subscription/service_layer/services.py
@@ -1,21 +1,16 @@
-import json
 import random
-from dataclasses import asdict
-from datetime import date, datetime, timedelta
+from datetime import date, datetime
 
 from tenacity import retry, stop_after_attempt, wait_fixed
 
 from subscription.domain.domain_models import (
-    Card,
     Payment,
     PaymentMethod,
     PaymentMethodType,
     PaymentStatus,
     SubscriptionPlan,
-    SubscriptionStatus,
     UserSubscription,
 )
-from subscription.serializers import CardSerializer
 from subscription.service_layer.unit_of_work import DjangoUnitOfWork
 
 
@@ -25,7 +20,7 @@ class SubscriptionService:
 
     # FIXME: 테스트 목적으로 임시로 추가한 메서드
     @retry(stop=stop_after_attempt(3), wait=wait_fixed(2))
-    def _process_payment(self, user_id: int, amount: float):
+    def process_payment(self, user_id: int, amount: float):
         payment_method = PaymentMethod(
             method_type=PaymentMethodType.CREDIT_CARD,
             details={
@@ -55,114 +50,7 @@ class SubscriptionService:
 
         return payment_success, payment
 
-    def subscribe_user_to_plan(
-        self, user_id: int, plan_name: str, payment_details: dict
-    ):
-        with self.uow:
-            try:
-                plan: SubscriptionPlan | None = self.uow.subscription_plans.get(
-                    plan_name
-                )
-            except ValueError as e:
-                raise e
-
-            user_subscription = plan.create_user_subscription(
-                user_id, date.today(), plan.duration_days
-            )
-            self.uow.user_subscriptions.add(user_subscription)
-
-            if payment_details["method_type"] == PaymentMethodType.CREDIT_CARD.value:
-                card_serializer = CardSerializer(
-                    data={
-                        "card_number": payment_details["card_number"],
-                        "card_expiry": payment_details["expiration_date"],
-                        "card_cvc": payment_details["cvc"],
-                    }
-                )
-
-                if card_serializer.is_valid(raise_exception=True):
-                    details = card_serializer.data
-            else:
-                # NOTE: 다른 결제 수단이 추가되면 이곳에 추가
-                details = "{}"
-
-            payment_method = PaymentMethod(
-                method_type=PaymentMethodType(payment_details["method_type"]),
-                details=json.dumps(details),
-            )
-            self.uow.payment_methods.add(payment_method)
-
-            payment = Payment(
-                subscription=user_subscription,
-                payment_method=payment_method,
-                amount=plan.price,
-                date=date.today(),
-                status=PaymentStatus.SUCCESS,
-            )
-            self.uow.payments.add(payment)
-
-            return {
-                "success": True,
-                "message": f"User {user_id} has subscribed to {plan_name} plan successfully.",
-            }
-
-    def cancel_subscription(
-        self,
-        user_id: int,
-    ):
-        with self.uow:
-            try:
-                user_subscription = (
-                    self.uow.user_subscriptions.get_active_subscription_by_user_id(
-                        user_id
-                    )
-                )
-            except ValueError as e:
-                raise e
-
-            user_subscription.status = SubscriptionStatus.CANCELED
-            self.uow.user_subscriptions.update(user_subscription)
-            return {
-                "success": True,
-                "message": f"User {user_id} has cancelled subscription successfully.",
-            }
-
-    def renew_subscription(self, user_id: int):
-        with self.uow:
-            current_subscription: UserSubscription = (
-                self.uow.user_subscriptions.get_active_subscription_by_user_id(user_id)
-            )
-            current_subscription.renew()
-
-            payment_success, payment_details = self._process_payment(
-                user_id, current_subscription.plan.price
-            )
-            if not payment_success:
-                return {
-                    "success": False,
-                    "message": "Failed to process payment for subscription renewal.",
-                }
-
-            if current_subscription:
-                current_subscription.status = SubscriptionStatus.EXPIRED
-                self.uow.user_subscriptions.update(current_subscription)
-
-            new_subscription = UserSubscription(
-                user_id=user_id,
-                plan=current_subscription.plan,
-                start_date=datetime.today(),
-                end_date=datetime.today()
-                + timedelta(days=current_subscription.plan.duration_days),
-                status=SubscriptionStatus.ACTIVE,
-            )
-            self.uow.user_subscriptions.add(new_subscription)
-
-            return {
-                "success": True,
-                "message": "Subscription renewed successfully.",
-            }
-
-    def _calculate_proration(
+    def calculate_proration(
         self, current_subscription: UserSubscription, new_plan: SubscriptionPlan
     ):
         remaining_days = (current_subscription.end_date - date.today()).days
@@ -179,47 +67,3 @@ class SubscriptionService:
         ) * remaining_days
 
         return prorated_amount, remaining_days
-
-    def change_subscription_plan(self, user_id: int, new_plan_name: str):
-        with self.uow:
-            try:
-                current_subscription = (
-                    self.uow.user_subscriptions.get_active_subscription_by_user_id(
-                        user_id
-                    )
-                )
-                new_plan = self.uow.subscription_plans.get(new_plan_name)
-
-                prorated_amount, remaining_days = self._calculate_proration(
-                    current_subscription, new_plan
-                )
-
-                payment_success, payment_details = self._process_payment(
-                    user_id, prorated_amount
-                )
-                if not payment_success:
-                    return {
-                        "success": False,
-                        "message": "Failed to process payment for plan change.",
-                    }
-
-                current_subscription.end_date = date.today()
-                current_subscription.status = SubscriptionStatus.CANCELED
-                self.uow.user_subscriptions.update(current_subscription)
-
-                new_subscription = UserSubscription(
-                    user_id=user_id,
-                    plan=new_plan,
-                    start_date=date.today(),
-                    end_date=date.today() + timedelta(days=remaining_days),
-                    status=SubscriptionStatus.ACTIVE,
-                )
-                self.uow.user_subscriptions.add(new_subscription)
-
-                return {
-                    "success": True,
-                    "message": "Subscription plan changed successfully.",
-                }
-
-            except ValueError as e:
-                raise e
